+++ CHANGELOG.md
@@ -7,6 +7,10 @@ uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- The remote trash dialog now has a search field that filters trashed notes by name.
+
 ## [0.4.0] - 2026-09-12
 
 ### Added